Force cookie name to "sid"
Previously, we would always look for an incoming cookie named "sid" even
though CookieOptions could set an arbitary cookie name. In addition, the
default name when no CookieOptions was provided was "", so we'd set a
nameless cookie then try to get one named "sid".

Unfortunately, warp::cookie::optional takes a &'static str for the
cookie name, so there's not a good way to use the cookie name we get
from the CookieOptions struct. Until that changes, best fix is to just
force it to "sid" everywhere. Then at least warp-sessions works out of
the box.